Fix harmless warning about size_t to int cast
The begin_ value here is never going to be out of int range, so just
cast it explicitly to avoid a warning about doing it implicitly.

+++ src/backends/postgresql/vector-into-type.cpp
@@ -79,7 +79,7 @@ void postgresql_vector_into_type_backend::post_fetch(bool gotData, indicator * i
 
         int const endRow = statement_.currentRow_ + statement_.rowsToConsume_;
 
-        for (int curRow = statement_.currentRow_, i = begin_;
+        for (int curRow = statement_.currentRow_, i = static_cast<int>(begin_);
              curRow != endRow; ++curRow, ++i)
         {
             // first, deal with indicators
